Ch1 problem with update method, Ch2 lost with queues

+++ 01_Cafe_Repository/MenuItem.cs
@@ -12,14 +12,14 @@ namespace _01_Cafe_Repository
         public int MealNumber { get; set; }
         public string MealName { get; set; }
         public string Description { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public string Ingredients { get; set; }
 
 
 
 
         public MenuItem() { }
-        public MenuItem(int mealNumber, string mealName, string description, double price, string ingredients)
+        public MenuItem(int mealNumber, string mealName, string description, decimal price, string ingredients)
         {
             MealNumber = mealNumber;
             MealName = mealName;
